refactor(plugins): follow the sidekickArea.panel commands from the SDK PR
