+++ bigbluebutton-html5/imports/ui/components/plugins-engine/ui-commands/handler.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import PluginChatUiCommandsHandler from './chat/handler';
 import PluginSidekickOptionsContainerUiCommandsHandler from './sidekick-options-container/handler';
+import PluginSidekickAreaUiCommandsHandler from './sidekick-area/handler';
 import PluginPresentationAreaUiCommandsHandler from './presentation/handler';
 import PluginUserStatusUiCommandsHandler from './user-status/handler';
 import PluginConferenceUiCommandsHandler from './conference/handler';
@@ -22,6 +23,7 @@ const PluginUiCommandsHandler = () => (
     <PluginCaptionsUiCommandsHandler />
     <PluginNavBarUiCommandsHandler />
     <PluginSidekickOptionsContainerUiCommandsHandler />
+    <PluginSidekickAreaUiCommandsHandler />
     <PluginPresentationAreaUiCommandsHandler />
     <PluginUserStatusUiCommandsHandler />
     <PluginConferenceUiCommandsHandler />
